feat: Add a generic action for adding issues to a board.
For WG boards it's handy to be able to say any issue on the WG repo goes
to the relevant board so we don't lose issues.  Add a re-usable workflow
and template to take advantage of a re-usable workflow that does this.

There are some off-the-shelf options but they mostly operate with a PAT
and not with an installed app so it would require creating a new user
for it rather than using the existing automation github app we have on
the `openedx` org.