docs: Add a CONTRIBUTING.md file.
Add a CONTRIBUTING.md file that is relevant to the whole Open edX
project.

I considered moving the content of the wiki to here, but there are other
useful and relevant links over there and it's much easier for the
community to update so I'm gonna leave the information there and just
point to it.