feat: Be able to support other python versions for make upgrade.
Make the version of python we use for running `make upgrade`
configurable by the workflow user.

We also drop the use of a matrix in the workflow since that's not
actually something we want to support.  The upgrade job should only run
in one python version for any given repo that implements it.  Running in
multiple python versions would just overwrite the same files with
potentially different values which could lead to confusion.