feat: add email alerts functionality in repo health job
Using this feature we can send explicit email alerts on failure and success of the repo health job.
For example: The receiving email address might be of opsgenie service responsible for creating alerts. Once the email
is received, the opsgenie service will create an alert and notify the responsible team.

Two input flags are available to manage email alerts:
- To enable email alerts on failure the value of input flag ENABLE_EMAIL_ALERTS should be set to true.
- To enable email alerts on success the value of both the input flags ENABLE_EMAIL_ALERTS and ENABLE_EMAIL_ALERTS_ON_SUCCESS should be set to true.

To send email alerts following configurations are required:
- EMAIL_ADDRESS_FOR_ALERTS
    Email address on which the alerts will be sent.
- SMTP_USERNAME
    Username of the email account from which the alerts will be sent.
- SMTP_PASSWORD
    Password of the email account from which the alerts will be sent.