feat: use pull request creator from repo-tools, not testeng-ci
testeng-ci is deprecated in favor of repo-tools [1]

This includes some cosmetic changes to the upgrade PRs.

Before:
* branch == jenkins/upgrade-python-requirements-{id}
* title  == Python Requirements Update
* commit == chore: Updating Python Requirements

After:
* branch == repo-tools/upgrade-python-requirements-{id}
* title  == chore: Upgrade Python requirements
* commit == chore: Upgrade Python requirements

NOTE: This leaves behind a testeng-ci reference in the repo-health job.
We need to figure out whether to fix the repo health dashboard or
deprecate it. This will a follow up task after [1] is closed.

[1] https://github.com/openedx/testeng-ci/issues/419