docs: Add more instructions to the Fast-Track DEPR template
Eventually, we update use the GitHub workflow to analyze the recognize
Fast-Track DEPR issues and apply the correct status & notification routing.
In the meantime, though, we're just adding more instructions here so
that contributors can follow the process themselves.

Additionally, we add some default values to the Dates fields
in order to make it clear what's going on in relation to the
standard DEPR process.
These values could be used in the future as keyphrases to trigger
the automation.